Move query tracker implementations to query-runner module
