+++ java/clients/src/main/java/sleeper/clients/status/report/TrackedQueryProgressReport.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2022-2023 Crown Copyright
+ * Copyright 2022-2024 Crown Copyright
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -22,8 +22,8 @@ import com.amazonaws.services.s3.AmazonS3ClientBuilder;
 
 import sleeper.clients.util.ClientUtils;
 import sleeper.configuration.properties.instance.InstanceProperties;
-import sleeper.query.tracker.DynamoDBQueryTracker;
-import sleeper.query.tracker.exception.QueryTrackerException;
+import sleeper.query.runner.tracker.DynamoDBQueryTracker;
+import sleeper.query.tracker.QueryTrackerException;
 
 /**
  *
