+++ java/query/query-core/src/main/java/sleeper/query/executor/QueryExecutor.java
@@ -29,9 +29,9 @@ import sleeper.core.range.Region;
 import sleeper.core.record.Record;
 import sleeper.core.statestore.StateStore;
 import sleeper.core.statestore.StateStoreException;
-import sleeper.query.QueryException;
 import sleeper.query.model.LeafPartitionQuery;
 import sleeper.query.model.Query;
+import sleeper.query.model.QueryException;
 import sleeper.query.recordretrieval.LeafPartitionQueryExecutor;
 import sleeper.query.recordretrieval.LeafPartitionRecordRetriever;
 import sleeper.query.recordretrieval.LeafPartitionRecordRetrieverImpl;
