+++ java/query/query-core/src/main/java/sleeper/query/tracker/QueryTrackerException.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2022-2023 Crown Copyright
+ * Copyright 2022-2024 Crown Copyright
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,19 +13,11 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package sleeper.query;
+package sleeper.query.tracker;
 
-public class QueryException extends Exception {
+public class QueryTrackerException extends Exception {
 
-    public QueryException(Throwable cause) {
-        super(cause);
-    }
-
-    public QueryException(String message, Throwable e) {
-        super(message, e);
-    }
-
-    public QueryException(String message) {
+    public QueryTrackerException(String message) {
         super(message);
     }
 }
