+++ java/query/query-core/src/main/java/sleeper/query/tracker/QueryTrackerStore.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2022-2023 Crown Copyright
+ * Copyright 2022-2024 Crown Copyright
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -15,8 +15,6 @@
  */
 
 package sleeper.query.tracker;
-
-import sleeper.query.tracker.exception.QueryTrackerException;
 
 import java.util.List;
