+++ java/query/query-core/src/test/java/sleeper/query/executor/QueryExecutorIT.java
@@ -46,9 +46,9 @@ import sleeper.core.statestore.FileReference;
 import sleeper.core.statestore.StateStore;
 import sleeper.core.statestore.StateStoreException;
 import sleeper.ingest.IngestFactory;
-import sleeper.query.QueryException;
 import sleeper.query.model.LeafPartitionQuery;
 import sleeper.query.model.Query;
+import sleeper.query.model.QueryException;
 import sleeper.query.model.QueryProcessingConfig;
 import sleeper.statestore.FixedStateStoreProvider;
