+++ java/query/query-core/src/test/java/sleeper/query/executor/QueryExecutorTest.java
@@ -39,8 +39,8 @@ import sleeper.core.statestore.FileReference;
 import sleeper.core.statestore.FileReferenceFactory;
 import sleeper.core.statestore.StateStore;
 import sleeper.core.statestore.StateStoreException;
-import sleeper.query.QueryException;
 import sleeper.query.model.Query;
+import sleeper.query.model.QueryException;
 import sleeper.query.model.QueryProcessingConfig;
 import sleeper.query.recordretrieval.InMemoryLeafPartitionRecordRetriever;
 import sleeper.query.recordretrieval.RecordRetrievalException;
