+++ java/query/query-lambda/src/main/java/sleeper/query/lambda/SqsLeafPartitionQueryProcessor.java
@@ -31,8 +31,8 @@ import sleeper.configuration.properties.table.TablePropertiesProvider;
 import sleeper.core.iterator.CloseableIterator;
 import sleeper.core.record.Record;
 import sleeper.io.parquet.utils.HadoopConfigurationProvider;
-import sleeper.query.QueryException;
 import sleeper.query.model.LeafPartitionQuery;
+import sleeper.query.model.QueryException;
 import sleeper.query.model.QueryOrLeafPartitionQuery;
 import sleeper.query.output.ResultsOutput;
 import sleeper.query.output.ResultsOutputConstants;
@@ -41,8 +41,8 @@ import sleeper.query.output.S3ResultsOutput;
 import sleeper.query.output.SQSResultsOutput;
 import sleeper.query.output.WebSocketResultsOutput;
 import sleeper.query.recordretrieval.LeafPartitionQueryExecutor;
+import sleeper.query.runner.tracker.DynamoDBQueryTracker;
 import sleeper.query.runner.tracker.QueryStatusReportListeners;
-import sleeper.query.tracker.DynamoDBQueryTracker;
 
 import java.io.IOException;
 import java.util.Collections;
