+++ java/query/query-lambda/src/main/java/sleeper/query/lambda/WebSocketQueryProcessorLambda.java
@@ -39,8 +39,8 @@ import sleeper.query.model.Query;
 import sleeper.query.model.QuerySerDe;
 import sleeper.query.output.ResultsOutputConstants;
 import sleeper.query.output.WebSocketResultsOutput;
+import sleeper.query.runner.tracker.WebSocketQueryStatusReportDestination;
 import sleeper.query.tracker.QueryStatusReportListener;
-import sleeper.query.tracker.WebSocketQueryStatusReportDestination;
 
 import java.nio.ByteBuffer;
 import java.nio.charset.StandardCharsets;
